+++ ui/src/features/AuthenticatedPage/index.tsx
@@ -1,10 +1,7 @@
 import { withAuthenticationRequired } from "@auth0/auth0-react";
 import { ReactNode } from "react";
 
-import { useAuthenticationRequired } from "./useAuth";
-
-export { AuthProvider } from "./authProvider";
-export { useAuth, useAuthenticationRequired } from "./useAuth";
+import { useAuthenticationRequired } from "@flow/lib/auth";
 
 const AuthenticationRequiredPage: React.FC<{ children?: ReactNode }> = ({ children }) => {
   const [isAuthenticated] = useAuthenticationRequired(); // TODO: show error
@@ -15,4 +12,4 @@ const withAuthorisation = (): ((props: any) => React.FC<any>) => {
   return withAuthenticationRequired as unknown as (props: any) => React.FC<any>;
 };
 
-export const AuthenticatedPage = withAuthorisation()(AuthenticationRequiredPage);
+export default withAuthorisation()(AuthenticationRequiredPage);
