+++ ui/src/lib/auth/authHook.ts
